requirements.txt
streamli
pandas
plotly
supabase
python-dotenv
requests